# docs/openapi.yaml
openapi: 3.1.0
info:
  title: substrate gateway
  version: 0.1.0
  summary: HTTP gateway for the substrate orchestration runtime.
  description: |
    The `psub-gateway` exposes a polyglot HTTP surface for substrate:

    - OpenAI-compatible chat completions (`/v1/chat/completions`, `/v1/models`)
    - A2A v0.3 message and task endpoints (`/a2a/*`)
    - Health, metrics, and admin surfaces for ops

    Authentication is API-key based (see `Authorization` below) for `/v1/*` and
    `/a2a/*`; admin routes additionally require `X-Admin-Token`.

    Wire format is JSON for request/response and `text/event-stream` for SSE
    streams (`stream=true` on chat completions).

    Source: `crates/psub-gateway/src/lib.rs` (router built at lines 257–298).
  contact:
    name: substrate maintainers
    url: https://github.com/<REDACTED>/substrate
  license:
    name: Apache-2.0 OR MIT
    url: https://github.com/<REDACTED>/substrate/blob/main/LICENSE-APACHE
servers:
  - url: http://127.0.0.1:8080
    description: Local dev (default `psub-gateway` bind)
  - url: https://gateway.example.invalid
    description: Production placeholder
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/chat/completions:
    post:
      tags: [chat]
      summary: Create a chat completion (OpenAI-compatible).
      description: |
        Accepts an OpenAI-compatible chat completion request and routes it
        through the configured provider chain. Set `stream=true` to receive
        an SSE stream of `ChatCompletionChunk` events instead of a single
        JSON response.

        See `crates/psub-gateway/src/openai.rs:35` for the request shape and
        `crates/psub-gateway/src/lib.rs:260` for the route binding.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatCompletionRequest' }
            examples:
              non_streaming:
                value:
                  model: claude-3-5-sonnet
                  messages:
                    - role: user
                      content: Hello, world.
                  stream: false
              streaming:
                value:
                  model: claude-3-5-sonnet
                  messages:
                    - role: user
                      content: Stream me a poem.
                  stream: true
      responses:
        '200':
          description: Completion (JSON) or stream (SSE) depending on `stream`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatCompletionResponse' }
            text/event-stream:
              schema: { $ref: '#/components/schemas/ChatCompletionChunk' }
        '400':
          description: Validation failure (empty messages, no user message, etc.).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          description: All upstream providers failed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Bearer-style API key. Format: `Authorization: Bearer <KEY>`.
        The header value is parsed from `x-session-id` for session affinity.
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Admin token from `SUBSTRATE_ADMIN_TOKEN` env var.
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    RateLimited:
      description: Rate limit exceeded for the current API key.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds until the limit window resets.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          description: Requested model id (may be overridden by routing).
        messages:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ChatMessage' }
        stream:
          type: boolean
          default: false
          description: When true, the response is a stream of SSE chunks.
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string
          minLength: 1
    ChatCompletionResponse:
      type: object
      required: [id, object, model, choices]
      properties:
        id: { type: string, format: uuid }
        object:
          type: string
          enum: [chat.completion]
        model:
          type: string
          description: Routed model id.
        choices:
          type: array
          items: { $ref: '#/components/schemas/ChatChoice' }
    ChatChoice:
      type: object
      required: [index, message, finish_reason]
      properties:
        index: { type: integer, minimum: 0 }
        message: { $ref: '#/components/schemas/ChatChoiceMessage' }
        finish_reason:
          type: string
          enum: [stop, length, content_filter, tool_calls]
    ChatChoiceMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [assistant]
        content: { type: string }
    ChatCompletionChunk:
      type: object
      description: |
        SSE chunk emitted when `stream=true`. Wrapped in
        `data: <json>\n\n` per SSE convention; ends with `data: [DONE]\n\n`.
      required: [id, object, model, choices]
      properties:
        id: { type: string, format: uuid }
        object:
          type: string
          enum: [chat.completion.chunk]
        model: { type: string }
        choices:
          type: array
          items: { $ref: '#/components/schemas/ChatChunkChoice' }
    ChatChunkChoice:
      type: object
      properties:
        index: { type: integer, minimum: 0 }
        delta:
          type: object
          properties:
            role: { type: string, enum: [assistant] }
            content: { type: string }
        finish_reason:
          type: string
          nullable: true
          enum: [stop, length, content_filter, tool_calls, null]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, message]
          properties:
            type:
              type: string
              enum:
                - invalid_request
                - authentication
                - rate_limit
                - upstream_failure
                - internal
            message: { type: string }
            code:
              type: string
              description: Optional machine-readable error code.